requirements.txt

viam-sdk[mlmodel]==0.51.0

typing-extensions
huggingface_hub
